Hip9011 and cleanups (#2477)
* hip9011: move hipCs out of EnginePins to hip9011 driver

* hip9011: cleanup debug output, save few RAM bytes

* Supress few signed vs unsigned warnings

+++ firmware/controllers/engine_cycle/main_trigger_callback.cpp
@@ -79,7 +79,7 @@ void startSimultaniousInjection(Engine *engine) {
 	EXPAND_Engine;
 #endif // EFI_UNIT_TEST
 	efitick_t nowNt = getTimeNowNt();
-	for (int i = 0; i < CONFIG(specs.cylindersCount); i++) {
+	for (size_t i = 0; i < CONFIG(specs.cylindersCount); i++) {
 		enginePins.injectors[i].open(nowNt PASS_ENGINE_PARAMETER_SUFFIX);
 	}
 }
@@ -89,7 +89,7 @@ static void endSimultaniousInjectionOnlyTogglePins(Engine *engine) {
 	EXPAND_Engine;
 #endif
 	efitick_t nowNt = getTimeNowNt();
-	for (int i = 0; i < CONFIG(specs.cylindersCount); i++) {
+	for (size_t i = 0; i < CONFIG(specs.cylindersCount); i++) {
 		enginePins.injectors[i].close(nowNt PASS_ENGINE_PARAMETER_SUFFIX);
 	}
 }
